requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
PyYAML>=6.0
